routes nearby page now displays nearby routes:

+++ app/controllers/welcome_controller.rb
@@ -3,8 +3,8 @@ class WelcomeController < ApplicationController
   end
 
   def nearby
-    @lat = params[:lat].to_i
-    @lng = params[:lng].to_i
+    @lat = params[:lat].to_f
+    @lng = params[:lng].to_f
     @tolerance = 0.031851 
     @routes = Route.where(
       "starting_lat > ? AND starting_lat < ?", @lat-@tolerance, @lat+@tolerance).where("starting_lng > ? AND starting_lng < ?", @lng-@tolerance, @lng+@tolerance)
